requirements_dev.txt
#Dev and build packages

# Testing
pytest==3.6.3
pytest-cov==2.5.1
pytest-instafail==0.4.0
pytest-watch==4.2.0
pytest-xdist==1.22.5
requests-mock==1.5.2

tox==3.1.2
